metrics: Drop the serialized request column
I had decided to keep that originally so that in case of a dramatic
failure we would be able to reparse everything from the raw bytes to fix
whatever issue had come up.

It seems like things are working just fine though, so the time has come
to remove it and reclaim some disk space.