Store machine location into metrics_machine table
Instead of storing machine location in events, we can store it in the
metrics_machine table and upsert when new events happen.

The deployment creates the new columns and sets the values according to the
events already stored in the database.

The event table is not removed yet. This way, we can check that everything is
OK for a while, and keep events if anything goes wrong.

This is the first step of https://phabricator.endlessm.com/T30471.